Fix Github issue #50, and fixes tab bar when rotated.
In React Native, the RCTTabarBar was changed to work better when the device was
rotated. Unfortunately, the fix was never applied over to this library.

The original issue is in facebook/react-native#1064. This commit applies the
changes from facebook/react-native@d82aaa7fb3a69c1ee736359ac53f3b8fd583b78b back in to this library.